refactor(tui): extract menu effect application
